==> 2-Selecting_initial_Realization.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
Realization_Feat=pd.DataFrame()
for items in ["02","04","50","56","58"]:
    ave_perm_KG="Ave_Perm_KG_"+items+".csv"
    ave_perm_phi_KG="Ave_Perm_Phi_KG_"+items+".csv"
    armse_KG="ARMSE_KG_"+items+".csv"
    armse_phi_KG="ARMSE_Phi_KG_"+items+".csv"
    std_perm_KG="STD_Perm_KG_"+items+".csv"
    std_perm_phi_KG="STD_Perm_Phi_KG_"+items+".csv"
    rmse_std_KG="RMSE_std_KG_"+items+".csv"
    rmse_std_phi_KG="RMSE_std_Phi_KG_"+items+".csv"
    
    Realization_Feat["Ave_Perm_KG_"+items]=pd.read_csv(ave_perm_KG,header=None)[0]
    Realization_Feat["Ave_Perm_Phi_KG_"+items]=pd.read_csv(ave_perm_phi_KG,header=None)[0]
    Realization_Feat["ARMSE_KG_"+items]=pd.read_csv(armse_KG,header=None)[0]
    Realization_Feat["ARMSE_Phi_KG_"+items]=pd.read_csv(armse_phi_KG,header=None)[0]
    Realization_Feat["STD_Perm_KG_"+items]=pd.read_csv(std_perm_KG,header=None)[0]
    Realization_Feat["STD_Perm_Phi_KG_"+items]=pd.read_csv(std_perm_phi_KG,header=None)[0]
    Realization_Feat["RMSE_std_KG_"+items]=pd.read_csv(rmse_std_KG,header=None)[0]
    Realization_Feat["RMSE_std_Phi_KG_"+items]=pd.read_csv(rmse_std_phi_KG,header=None)[0]
    
    
    
Realization_Feat["variog"]=np.zeros([400,1])
for i in range(0,100):
    Realization_Feat["variog"][i]=1
    Realization_Feat["variog"][i+101]=2
    Realization_Feat["variog"][i+201]=3
    Realization_Feat["variog"][i+300]=4
    Realization_Feat["variog"][100]=1
    Realization_Feat["variog"][300]=3

Realization_Feat.to_csv("Realization_Feat.csv")
                                       

normalized_real_feat=(Realization_Feat-Realization_Feat.min())/(Realization_Feat.max()-Realization_Feat.min())
from sklearn.cluster import KMeans
Init_Realization_Feat1=normalized_real_feat.loc[list(range(0,100))]
Init_Realization_Feat2=normalized_real_feat.loc[list(range(100,200))]
Init_Realization_Feat3=normalized_real_feat.loc[list(range(200,300))]
Init_Realization_Feat4=normalized_real_feat.loc[list(range(300,400))]
kmeans=KMeans(n_clusters=15)
kmeans.fit(Init_Realization_Feat1)
a1=kmeans.labels_
b1=kmeans.get_params
kmeans.fit(Init_Realization_Feat2)
a2=kmeans.labels_
b2=kmeans.get_params
kmeans.fit(Init_Realization_Feat3)
a3=kmeans.labels_
b3=kmeans.get_params
kmeans.fit(Init_Realization_Feat4)
a4=kmeans.labels_
b4=kmeans.get_params

Init_Realization_Feat1["Group"]=a1
Init_Realization_Feat2["Group"]=a2
Init_Realization_Feat3["Group"]=a3
Init_Realization_Feat4["Group"]=a4

a=[]
for cluster in range(0,15):
    real_num=np.array(Init_Realization_Feat1[Init_Realization_Feat1["Group"]==cluster].index)
    X_trans_cluster=np.array(Init_Realization_Feat1.loc[real_num])
    dist=np.ones([1,X_trans_cluster.shape[0]])
    for items in range(0,X_trans_cluster.shape[0]):
        for realization in range(0,X_trans_cluster.shape[0]):
            dist[0,items]+=np.sqrt(((X_trans_cluster[items]-X_trans_cluster[realization])**2).sum())
    a.append(real_num[dist.argmin()])
for cluster in range(0,15):
    real_num=np.array(Init_Realization_Feat2[Init_Realization_Feat2["Group"]==cluster].index)
    X_trans_cluster=np.array(Init_Realization_Feat2.loc[real_num])
    dist=np.ones([1,X_trans_cluster.shape[0]])
    for items in range(0,X_trans_cluster.shape[0]):
        for realization in range(0,X_trans_cluster.shape[0]):
            dist[0,items]+=np.sqrt(((X_trans_cluster[items]-X_trans_cluster[realization])**2).sum())
    a.append(real_num[dist.argmin()])
for cluster in range(0,15):
    real_num=np.array(Init_Realization_Feat3[Init_Realization_Feat3["Group"]==cluster].index)
    X_trans_cluster=np.array(Init_Realization_Feat3.loc[real_num])
    dist=np.ones([1,X_trans_cluster.shape[0]])
    for items in range(0,X_trans_cluster.shape[0]):
        for realization in range(0,X_trans_cluster.shape[0]):
            dist[0,items]+=np.sqrt(((X_trans_cluster[items]-X_trans_cluster[realization])**2).sum())
    a.append(real_num[dist.argmin()])
for cluster in range(0,15):
    real_num=np.array(Init_Realization_Feat4[Init_Realization_Feat4["Group"]==cluster].index)
    X_trans_cluster=np.array(Init_Realization_Feat4.loc[real_num])
    dist=np.ones([1,X_trans_cluster.shape[0]])
    for items in range(0,X_trans_cluster.shape[0]):
        for realization in range(0,X_trans_cluster.shape[0]):
            dist[0,items]+=np.sqrt(((X_trans_cluster[items]-X_trans_cluster[realization])**2).sum())
    a.append(real_num[dist.argmin()])
np.savetxt("Initial_Real_For_ML" , a , delimiter =",", fmt ='% .0f')


import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
Realization_Feat=pd.DataFrame()
for items in ["02","04","50","56","58"]:
    ave_perm_KG="Ave_Perm_KG_"+items+".csv"
    ave_perm_phi_KG="Ave_Perm_Phi_KG_"+items+".csv"
    armse_KG="ARMSE_KG_"+items+".csv"
    armse_phi_KG="ARMSE_Phi_KG_"+items+".csv"
    std_perm_KG="STD_Perm_KG_"+items+".csv"
    std_perm_phi_KG="STD_Perm_Phi_KG_"+items+".csv"
    rmse_std_KG="RMSE_std_KG_"+items+".csv"
    rmse_std_phi_KG="RMSE_std_Phi_KG_"+items+".csv"
    
    Realization_Feat["Ave_Perm_KG_"+items]=pd.read_csv(ave_perm_KG,header=None)[0]
    Realization_Feat["Ave_Perm_Phi_KG_"+items]=pd.read_csv(ave_perm_phi_KG,header=None)[0]
    Realization_Feat["ARMSE_KG_"+items]=pd.read_csv(armse_KG,header=None)[0]
    Realization_Feat["ARMSE_Phi_KG_"+items]=pd.read_csv(armse_phi_KG,header=None)[0]
    Realization_Feat["STD_Perm_KG_"+items]=pd.read_csv(std_perm_KG,header=None)[0]
    Realization_Feat["STD_Perm_Phi_KG_"+items]=pd.read_csv(std_perm_phi_KG,header=None)[0]
    Realization_Feat["RMSE_std_KG_"+items]=pd.read_csv(rmse_std_KG,header=None)[0]
    Realization_Feat["RMSE_std_Phi_KG_"+items]=pd.read_csv(rmse_std_phi_KG,header=None)[0]
