==> 3-Extracting_the_Wells_Behaviors_from_RMS_files.ipynb <==
import pandas as pd
import numpy as np
Real_M=list(pd.read_csv("Initial_Real_For_ML", header=None)[0])
Real=list(np.array(Real_M)+1)
Real_ML=["HM_200"]+Real
Days=list(pd.read_csv("Days", header=None)[0])


## Extracting the Wells Behavior
indx=[]
indx_BP9=[]
indx_GPR=[]
WOPR_KG_02=[]
WOPR_KG_04=[]
WOPR_KG_50=[]
WOPR_KG_56=[]
WOPR_KG_58=[]

WGPR_KG_02=[]
WGPR_KG_04=[]
WGPR_KG_50=[]
WGPR_KG_56=[]
WGPR_KG_58=[]

WBP9_KG_02=[]
WBP9_KG_04=[]
WBP9_KG_50=[]
WBP9_KG_56=[]
WBP9_KG_58=[]


m=0
for item in Real_ML:
    name=[str(item),".RSM"]
    names="".join(name)
    a=pd.read_table(names, header=None)

    w=(a[a[4]=='WOPR        '].index.tolist())
    index1=[]
    for i in Days:
        index1.append(a[a[1]==str(i)].index.tolist())
    for j in range(0,len(Days)):
        indx.append(index1[j][10])
        indx_BP9.append(index1[j][1])
        indx_GPR.append(index1[j][11])   
    for k in indx[len(Days)*m:len(Days)*m+len(Days)]:
        WOPR_KG_02.append(float(a[3][k]))
        WOPR_KG_04.append(float(a[4][k]))
        WOPR_KG_50.append(float(a[5][k]))
        WOPR_KG_56.append(float(a[6][k]))
        WOPR_KG_58.append(float(a[7][k]))
    for q in indx_BP9[len(Days)*m:len(Days)*m+len(Days)]:
        WBP9_KG_02.append(float(a[6][q]))
        WBP9_KG_04.append(float(a[7][q]))
        WBP9_KG_50.append(float(a[8][q]))
        WBP9_KG_56.append(float(a[9][q]))
        WBP9_KG_58.append(float(a[10][q]))
    for s in indx_GPR[len(Days)*m:len(Days)*m+len(Days)]:
        WGPR_KG_02.append(float(a[2][s]))
        WGPR_KG_04.append(float(a[3][s]))
        WGPR_KG_50.append(float(a[4][s]))
        WGPR_KG_56.append(float(a[5][s]))
        WGPR_KG_58.append(float(a[6][s]))
    m+=1


WOPR_KG_02=np.transpose(np.reshape(WOPR_KG_02,(61,len(Days))))
WOPR_KG_04=np.transpose(np.reshape(WOPR_KG_04,(61,len(Days))))
WOPR_KG_50=np.transpose(np.reshape(WOPR_KG_50,(61,len(Days))))
WOPR_KG_56=np.transpose(np.reshape(WOPR_KG_56,(61,len(Days))))
WOPR_KG_58=np.transpose(np.reshape(WOPR_KG_58,(61,len(Days))))

WGPR_KG_02=np.transpose(np.reshape(WGPR_KG_02,(61,len(Days))))
WGPR_KG_04=np.transpose(np.reshape(WGPR_KG_04,(61,len(Days))))
WGPR_KG_50=np.transpose(np.reshape(WGPR_KG_50,(61,len(Days))))
WGPR_KG_56=np.transpose(np.reshape(WGPR_KG_56,(61,len(Days))))
WGPR_KG_58=np.transpose(np.reshape(WGPR_KG_58,(61,len(Days))))


WBP9_KG_02=np.transpose(np.reshape(WBP9_KG_02,(61,len(Days))))
WBP9_KG_04=np.transpose(np.reshape(WBP9_KG_04,(61,len(Days))))
WBP9_KG_50=np.transpose(np.reshape(WBP9_KG_50,(61,len(Days))))
WBP9_KG_56=np.transpose(np.reshape(WBP9_KG_56,(61,len(Days))))
WBP9_KG_58=np.transpose(np.reshape(WBP9_KG_58,(61,len(Days))))


np.savetxt("WOPR_KG_02.csv" , WOPR_KG_02 , delimiter =", ", fmt ='% s') 
np.savetxt("WOPR_KG_04.csv" , WOPR_KG_04 , delimiter =", ", fmt ='% s') 
np.savetxt("WOPR_KG_50.csv" , WOPR_KG_50 , delimiter =", ", fmt ='% s') 
np.savetxt("WOPR_KG_56.csv" , WOPR_KG_56 , delimiter =", ", fmt ='% s') 
np.savetxt("WOPR_KG_58.csv" , WOPR_KG_58 , delimiter =", ", fmt ='% s') 


np.savetxt("WGPR_KG_02.csv" , WGPR_KG_02 , delimiter =", ", fmt ='% s') 
np.savetxt("WGPR_KG_04.csv" , WGPR_KG_04 , delimiter =", ", fmt ='% s') 
np.savetxt("WGPR_KG_50.csv" , WGPR_KG_50 , delimiter =", ", fmt ='% s') 
np.savetxt("WGPR_KG_56.csv" , WGPR_KG_56 , delimiter =", ", fmt ='% s') 
np.savetxt("WGPR_KG_58.csv" , WGPR_KG_58 , delimiter =", ", fmt ='% s') 


np.savetxt("WBP9_KG_02.csv" , WBP9_KG_02 , delimiter =", ", fmt ='% s') 
np.savetxt("WBP9_KG_04.csv" , WBP9_KG_04 , delimiter =", ", fmt ='% s') 
np.savetxt("WBP9_KG_50.csv" , WBP9_KG_50 , delimiter =", ", fmt ='% s')
np.savetxt("WBP9_KG_56.csv" , WBP9_KG_56 , delimiter =", ", fmt ='% s') 
np.savetxt("WBP9_KG_58.csv" , WBP9_KG_58 , delimiter =", ", fmt ='% s')
