==> 4-Extract_the_Diffs.ipynb <==
import pandas as pd
import numpy as np
Real_M=list(pd.read_csv("Initial_Real_For_ML", header=None)[0])
Real=list(np.array(Real_M)+1)
Real_ML=["HM_200"]+Real

WOPR_KG_02=pd.read_table("WOPR_KG_02.csv" , header=None , delimiter=", ")
WOPR_KG_04=pd.read_table("WOPR_KG_04.csv" , header=None , delimiter=", ")
WOPR_KG_50=pd.read_table("WOPR_KG_50.csv" , header=None , delimiter=", ")
WOPR_KG_56=pd.read_table("WOPR_KG_56.csv" , header=None , delimiter=", ")
WOPR_KG_58=pd.read_table("WOPR_KG_58.csv" , header=None , delimiter=", ")

WGPR_KG_02=pd.read_table("WGPR_KG_02.csv" , header=None , delimiter=", ")
WGPR_KG_04=pd.read_table("WGPR_KG_04.csv" , header=None , delimiter=", ")
WGPR_KG_50=pd.read_table("WGPR_KG_50.csv" , header=None , delimiter=", ")
WGPR_KG_56=pd.read_table("WGPR_KG_56.csv" , header=None , delimiter=", ")
WGPR_KG_58=pd.read_table("WGPR_KG_58.csv" , header=None , delimiter=", ")

WBP9_KG_02=pd.read_table("WBP9_KG_02.csv" , header=None , delimiter=", ")
WBP9_KG_04=pd.read_table("WBP9_KG_04.csv" , header=None , delimiter=", ")
WBP9_KG_50=pd.read_table("WBP9_KG_50.csv" , header=None , delimiter=", ")
WBP9_KG_56=pd.read_table("WBP9_KG_56.csv" , header=None , delimiter=", ")
WBP9_KG_58=pd.read_table("WBP9_KG_58.csv" , header=None , delimiter=", ")


WOPR_KG_02_diff=np.zeros((61,550))
WOPR_KG_04_diff=np.zeros((61,550))
WOPR_KG_50_diff=np.zeros((61,550))
WOPR_KG_56_diff=np.zeros((61,550))
WOPR_KG_58_diff=np.zeros((61,550))

WGPR_KG_02_diff=np.zeros((61,550))
WGPR_KG_04_diff=np.zeros((61,550))
WGPR_KG_50_diff=np.zeros((61,550))
WGPR_KG_56_diff=np.zeros((61,550))
WGPR_KG_58_diff=np.zeros((61,550))

WBP9_KG_02_diff=np.zeros((61,550))
WBP9_KG_04_diff=np.zeros((61,550))
WBP9_KG_50_diff=np.zeros((61,550))
WBP9_KG_56_diff=np.zeros((61,550))
WBP9_KG_58_diff=np.zeros((61,550))
for k in range(0,61):
    WOPR_KG_02[k].replace(to_replace = 0, value = 0.00001, inplace=True)
    WOPR_KG_04[k].replace(to_replace = 0, value = 0.00001, inplace=True)
    WOPR_KG_50[k].replace(to_replace = 0, value = 0.00001, inplace=True)
    WOPR_KG_56[k].replace(to_replace = 0, value = 0.00001, inplace=True)
    WOPR_KG_58[k].replace(to_replace = 0, value = 0.00001, inplace=True)
    
    WGPR_KG_02[k].replace(to_replace = 0, value = 0.00001, inplace=True)
    WGPR_KG_04[k].replace(to_replace = 0, value = 0.00001, inplace=True)
    WGPR_KG_50[k].replace(to_replace = 0, value = 0.00001, inplace=True)
    WGPR_KG_56[k].replace(to_replace = 0, value = 0.00001, inplace=True)
    WGPR_KG_58[k].replace(to_replace = 0, value = 0.00001, inplace=True)
    
    WBP9_KG_02[k].replace(to_replace = 0, value = 0.00001, inplace=True)
    WBP9_KG_04[k].replace(to_replace = 0, value = 0.00001, inplace=True)
    WBP9_KG_50[k].replace(to_replace = 0, value = 0.00001, inplace=True)
    WBP9_KG_56[k].replace(to_replace = 0, value = 0.00001, inplace=True)
    WBP9_KG_58[k].replace(to_replace = 0, value = 0.00001, inplace=True)     
    
    ################################################################
    #OIL PRODUCTION RATE DIFF.
    ################################################################
    m=0
    for items in WOPR_KG_02[0]:
        WOPR_KG_02_diff[k][m]=((WOPR_KG_02[k][m]/WOPR_KG_02[0][m])-1)  
        m+=1
    m=0
    for items in WOPR_KG_04[0]:
        WOPR_KG_04_diff[k][m]=((WOPR_KG_04[k][m]/WOPR_KG_04[0][m])-1)  
        m+=1
    m=0
    for items in WOPR_KG_50[0]:
        WOPR_KG_50_diff[k][m]=((WOPR_KG_50[k][m]/WOPR_KG_50[0][m])-1)  
        m+=1
    m=0
    for items in WOPR_KG_56[0]:
        WOPR_KG_56_diff[k][m]=((WOPR_KG_56[k][m]/WOPR_KG_56[0][m])-1)  
        m+=1
    m=0
    for items in WOPR_KG_58[0]:
        WOPR_KG_58_diff[k][m]=((WOPR_KG_58[k][m]/WOPR_KG_58[0][m])-1)  
        m+=1
    
    ################################################################
    #GAS PRODUCTION RATE DIFF.
    ################################################################
    m=0
    for items in WGPR_KG_02[0]:
        WGPR_KG_02_diff[k][m]=((WGPR_KG_02[k][m]/WGPR_KG_02[0][m])-1)  
        m+=1
    m=0
    for items in WGPR_KG_04[0]:
        WGPR_KG_04_diff[k][m]=((WGPR_KG_04[k][m]/WGPR_KG_04[0][m])-1)  
        m+=1
    m=0
    for items in WGPR_KG_50[0]:
        WGPR_KG_50_diff[k][m]=((WGPR_KG_50[k][m]/WGPR_KG_50[0][m])-1)  
        m+=1
    m=0
    for items in WGPR_KG_56[0]:
        WGPR_KG_56_diff[k][m]=((WGPR_KG_56[k][m]/WGPR_KG_56[0][m])-1)  
        m+=1
    m=0
    for items in WGPR_KG_58[0]:
        WGPR_KG_58_diff[k][m]=((WGPR_KG_58[k][m]/WGPR_KG_58[0][m])-1)  
        m+=1
    ################################################################
    #WELL PRESSURE DIFF.
    ################################################################
    m=0
    for items in WBP9_KG_02[0]:
        WBP9_KG_02_diff[k][m]=((WBP9_KG_02[k][m]/WBP9_KG_02[0][m])-1)  
        m+=1
    m=0
    for items in WBP9_KG_04[0]:
        WBP9_KG_04_diff[k][m]=((WBP9_KG_04[k][m]/WBP9_KG_04[0][m])-1)  
        m+=1
    m=0
    for items in WBP9_KG_50[0]:
        WBP9_KG_50_diff[k][m]=((WBP9_KG_50[k][m]/WBP9_KG_50[0][m])-1)  
        m+=1
    m=0
    for items in WBP9_KG_56[0]:
        WBP9_KG_56_diff[k][m]=((WBP9_KG_56[k][m]/WBP9_KG_56[0][m])-1)  
        m+=1
    m=0
    for items in WBP9_KG_58[0]:
        WBP9_KG_58_diff[k][m]=((WBP9_KG_58[k][m]/WBP9_KG_58[0][m])-1)  
        m+=1

#
WOPR_KG_02_diff=np.transpose(WOPR_KG_02_diff)
WOPR_KG_04_diff=np.transpose(WOPR_KG_04_diff)
WOPR_KG_50_diff=np.transpose(WOPR_KG_50_diff)
WOPR_KG_56_diff=np.transpose(WOPR_KG_56_diff)
WOPR_KG_58_diff=np.transpose(WOPR_KG_58_diff)

WGPR_KG_02_diff=np.transpose(WGPR_KG_02_diff)
WGPR_KG_04_diff=np.transpose(WGPR_KG_04_diff)
WGPR_KG_50_diff=np.transpose(WGPR_KG_50_diff)
WGPR_KG_56_diff=np.transpose(WGPR_KG_56_diff)
WGPR_KG_58_diff=np.transpose(WGPR_KG_58_diff)

WBP9_KG_02_diff=np.transpose(WBP9_KG_02_diff)
WBP9_KG_04_diff=np.transpose(WBP9_KG_04_diff)
WBP9_KG_50_diff=np.transpose(WBP9_KG_50_diff)
WBP9_KG_56_diff=np.transpose(WBP9_KG_56_diff)
WBP9_KG_58_diff=np.transpose(WBP9_KG_58_diff)


np.savetxt("WOPR_KG_02_Diff.csv" , WOPR_KG_02_diff , delimiter =", ", fmt ='% s') 
np.savetxt("WOPR_KG_04_Diff.csv" , WOPR_KG_04_diff , delimiter =", ", fmt ='% s') 
np.savetxt("WOPR_KG_50_Diff.csv" , WOPR_KG_50_diff , delimiter =", ", fmt ='% s') 
np.savetxt("WOPR_KG_56_Diff.csv" , WOPR_KG_56_diff , delimiter =", ", fmt ='% s') 
np.savetxt("WOPR_KG_58_Diff.csv" , WOPR_KG_58_diff , delimiter =", ", fmt ='% s') 

np.savetxt("WGPR_KG_02_Diff.csv" , WGPR_KG_02_diff , delimiter =", ", fmt ='% s') 
np.savetxt("WGPR_KG_04_Diff.csv" , WGPR_KG_04_diff , delimiter =", ", fmt ='% s') 
np.savetxt("WGPR_KG_50_Diff.csv" , WGPR_KG_50_diff , delimiter =", ", fmt ='% s') 
np.savetxt("WGPR_KG_56_Diff.csv" , WGPR_KG_56_diff , delimiter =", ", fmt ='% s') 
np.savetxt("WGPR_KG_58_Diff.csv" , WGPR_KG_58_diff , delimiter =", ", fmt ='% s')  

np.savetxt("WBP9_KG_02_Diff.csv" , WBP9_KG_02_diff , delimiter =", ", fmt ='% s') 
np.savetxt("WBP9_KG_04_Diff.csv" , WBP9_KG_04_diff , delimiter =", ", fmt ='% s') 
np.savetxt("WBP9_KG_50_Diff.csv" , WBP9_KG_50_diff , delimiter =", ", fmt ='% s') 
np.savetxt("WBP9_KG_56_Diff.csv" , WBP9_KG_56_diff , delimiter =", ", fmt ='% s') 
np.savetxt("WBP9_KG_58_Diff.csv" , WBP9_KG_58_diff , delimiter =", ", fmt ='% s')
