==> 5-Realization_Selection_HM.ipynb <==
import pandas as pd
import numpy as np

WOPR_KG_02_mat=pd.read_csv("WOPR_KG_02_Diff.csv", header=None)
WOPR_KG_04_mat=pd.read_csv("WOPR_KG_04_Diff.csv", header=None)
WOPR_KG_50_mat=pd.read_csv("WOPR_KG_50_Diff.csv", header=None)
WOPR_KG_56_mat=pd.read_csv("WOPR_KG_56_Diff.csv", header=None)
WOPR_KG_58_mat=pd.read_csv("WOPR_KG_58_Diff.csv", header=None)

WGPR_KG_02_mat=pd.read_csv("WGPR_KG_02_Diff.csv", header=None)
WGPR_KG_04_mat=pd.read_csv("WGPR_KG_04_Diff.csv", header=None)
WGPR_KG_50_mat=pd.read_csv("WGPR_KG_50_Diff.csv", header=None)
WGPR_KG_56_mat=pd.read_csv("WGPR_KG_56_Diff.csv", header=None)
WGPR_KG_58_mat=pd.read_csv("WGPR_KG_58_Diff.csv", header=None)

WBP9_KG_02_mat=pd.read_csv("WBP9_KG_02_Diff.csv", header=None)
WBP9_KG_04_mat=pd.read_csv("WBP9_KG_04_Diff.csv", header=None)
WBP9_KG_50_mat=pd.read_csv("WBP9_KG_50_Diff.csv", header=None)
WBP9_KG_56_mat=pd.read_csv("WBP9_KG_56_Diff.csv", header=None)
WBP9_KG_58_mat=pd.read_csv("WBP9_KG_58_Diff.csv", header=None)

a=(WOPR_KG_02_mat.abs().sum())/550
b=(WOPR_KG_04_mat.abs().sum())/550
c=(WOPR_KG_50_mat.abs().sum())/550
d=(WOPR_KG_56_mat.abs().sum())/550
e=(WOPR_KG_58_mat.abs().sum())/550

f=(WGPR_KG_02_mat.abs().sum())/550
g=(WGPR_KG_04_mat.abs().sum())/550
h=(WGPR_KG_50_mat.abs().sum())/550
i=(WGPR_KG_56_mat.abs().sum())/550
j=(WGPR_KG_58_mat.abs().sum())/550

k=(WBP9_KG_02_mat.abs().sum())/550
l=(WBP9_KG_04_mat.abs().sum())/550
m=(WBP9_KG_50_mat.abs().sum())/550
n=(WBP9_KG_56_mat.abs().sum())/550
o=(WBP9_KG_58_mat.abs().sum())/550

a1=list(a[a<=0.03].index)
b1=list(b[b<=0.03].index)
c1=list(c[c<=0.03].index)
d1=list(d[d<=0.03].index)
e1=list(e[e<=0.03].index)
f1=list(f[f<=0.03].index)
g1=list(g[g<=0.03].index)
h1=list(h[h<=0.03].index)
i1=list(i[i<=0.03].index)
j1=list(j[j<=0.03].index)
k1=list(k[k<=0.03].index)
l1=list(l[l<=0.03].index)
m1=list(m[m<=0.03].index)
n1=list(n[n<=0.03].index)
o1=list(o[o<=0.03].index)

len(a1) , len(b1) , len(c1) , len(d1) , len(e1) , len(f1) , len(g1) , len(h1) , len(i1) , len(j1) , len(k1) , len(l1) , len(m1) , len(n1) , len(o1) 

v=[]
realization_HM=[]
for items in k1:
    m=0
    m+=(a1.count(items))
    m+=(b1.count(items))
    m+=(c1.count(items))
    m+=(d1.count(items))
    m+=(e1.count(items))
    m+=(f1.count(items))
    m+=(g1.count(items))
    m+=(i1.count(items))
    m+=(j1.count(items))
    m+=(h1.count(items))
    m+=(l1.count(items))
    m+=(m1.count(items))
    m+=(n1.count(items))
    m+=(o1.count(items))
    #print(m)
    v.append(m)
    if m==14:
        realization_HM.append(items)
print(v.count(14))
print(realization_HM)

np.savetxt("Realization_HM.csv" , realization_HM , delimiter =", ", fmt ='% s') 

Initial_Realizations=pd.read_csv("Initial_Real_For_ML", header=None)
Initial_Realizations["HM"]=0
for items in realization_HM:
    if items>0:
        Initial_Realizations["HM"][items-1]=1
np.savetxt("Initial_Realization_HM.csv" , Initial_Realizations , delimiter =", ", fmt ='% .0f')
