==> 8-Grid_Properties_Between_Inj_Prod.ipynb <==
import numpy as np
import pandas as pd
#Grid_btw_wells=pd.read_excel("grids_between_prod_inj.xlsx", header=0)
Well_Grids_hist=pd.read_excel("grids_between_prod_inj.xlsx", header=0)
KG_02_Grids_hist=Well_Grids_hist[["KG-02I_I","KG-02I_J"]]
KG_04_Grids_hist=Well_Grids_hist[["KG-04I_I","KG-04I_J"]]
KG_50_Grids_hist=Well_Grids_hist[["KG-50I_I","KG-50I_J"]]
KG_56_Grids_hist=Well_Grids_hist[["KG-56I_I","KG-56I_J"]]
KG_58_Grids_hist=Well_Grids_hist[["KG-58I_I","KG-58I_J"]]
KG_50_02_Grids_hist=Well_Grids_hist[["KG-50-02_I","KG-50-02_J"]]
KG_56_02_Grids_hist=Well_Grids_hist[["KG-56-02_I","KG-56-02_J"]]

KG_02_Grids_hist.dropna(inplace=True)
KG_04_Grids_hist.dropna(inplace=True)
KG_50_Grids_hist.dropna(inplace=True)
KG_56_Grids_hist.dropna(inplace=True)
KG_58_Grids_hist.dropna(inplace=True)
KG_50_02_Grids_hist.dropna(inplace=True)
KG_56_02_Grids_hist.dropna(inplace=True)


Ave_Perm_KG_02=[]
Ave_Perm_KG_04=[]
Ave_Perm_KG_50=[]
Ave_Perm_KG_56=[]
Ave_Perm_KG_58=[]
Ave_Perm_KG_50_02=[]
Ave_Perm_KG_56_02=[]

Ave_Perm_Phi_KG_02=[]
Ave_Perm_Phi_KG_04=[]
Ave_Perm_Phi_KG_50=[]
Ave_Perm_Phi_KG_56=[]
Ave_Perm_Phi_KG_58=[]
Ave_Perm_Phi_KG_50_02=[]
Ave_Perm_Phi_KG_56_02=[]

Ave_STD_KG_02=[]
Ave_STD_KG_04=[]
Ave_STD_KG_50=[]
Ave_STD_KG_56=[]
Ave_STD_KG_58=[]
Ave_STD_KG_50_02=[]
Ave_STD_KG_56_02=[]

Ave_STD_Phi_KG_02=[]
Ave_STD_Phi_KG_04=[]
Ave_STD_Phi_KG_50=[]
Ave_STD_Phi_KG_56=[]
Ave_STD_Phi_KG_58=[]
Ave_STD_Phi_KG_50_02=[]
Ave_STD_Phi_KG_56_02=[]



Poro=pd.read_csv("sector_porosity.csv",header=0, delimiter=",", index_col=None)
Perm=pd.read_csv("sector_perm.csv",header=0, delimiter=",", index_col=None)
a=list(Poro[Poro["Poro"]==0].index)
Poro.drop(a,axis=0,inplace=True)
Perm.drop(a,axis=0,inplace=True)

for name in range(1,401):
    
    Real1=Perm[["i","j","k",str(name)]]
    perm_KG_02=np.array([])
    perm_KG_04=np.array([])
    perm_KG_50=np.array([])
    perm_KG_56=np.array([])
    perm_KG_58=np.array([])
    perm_KG_50_02=np.array([])
    perm_KG_56_02=np.array([])

    perm_poro_KG_02=np.array([])
    perm_poro_KG_04=np.array([])
    perm_poro_KG_50=np.array([])
    perm_poro_KG_56=np.array([])
    perm_poro_KG_58=np.array([])
    perm_poro_KG_50_02=np.array([])
    perm_poro_KG_56_02=np.array([]) 
    
    for i in range(0,len(KG_02_Grids_hist.index)):
        perm_KG_02=np.r_[perm_KG_02,np.array(Real1.loc[(Real1['i'] == KG_02_Grids_hist["KG-02I_I"][i]) & (Real1['j'] == KG_02_Grids_hist["KG-02I_J"][i])][str(name)])]
        perm_poro_KG_02=np.r_[perm_poro_KG_02,np.array(Real1.loc[(Real1['i'] == KG_02_Grids_hist["KG-02I_I"][i]) & (Real1['j'] == KG_02_Grids_hist["KG-02I_J"][i])][str(name)])/np.array(Poro.loc[(Poro['i'] == KG_02_Grids_hist["KG-02I_I"][i]) & (Poro['j'] == KG_02_Grids_hist["KG-02I_J"][i])]["Poro"])]
    
    for i in range(0,len(KG_04_Grids_hist.index)):
        perm_KG_04=np.r_[perm_KG_04,np.array(Real1.loc[(Real1['i'] == KG_04_Grids_hist["KG-04I_I"][i]) & (Real1['j'] == KG_04_Grids_hist["KG-04I_J"][i])][str(name)])]
        perm_poro_KG_04=np.r_[perm_poro_KG_04,np.array(Real1.loc[(Real1['i'] == KG_04_Grids_hist["KG-04I_I"][i]) & (Real1['j'] == KG_04_Grids_hist["KG-04I_J"][i])][str(name)])/np.array(Poro.loc[(Poro['i'] == KG_04_Grids_hist["KG-04I_I"][i]) & (Poro['j'] == KG_04_Grids_hist["KG-04I_J"][i])]["Poro"])]

              
    for i in range(0,len(KG_50_Grids_hist.index)):
        perm_KG_50=np.r_[perm_KG_50,np.array(Real1.loc[(Real1['i'] == KG_50_Grids_hist["KG-50I_I"][i]) & (Real1['j'] == KG_50_Grids_hist["KG-50I_J"][i])][str(name)])]
        perm_poro_KG_50=np.r_[perm_poro_KG_50,np.array(Real1.loc[(Real1['i'] == KG_50_Grids_hist["KG-50I_I"][i]) & (Real1['j'] == KG_50_Grids_hist["KG-50I_J"][i])][str(name)])/np.array(Poro.loc[(Poro['i'] == KG_50_Grids_hist["KG-50I_I"][i]) & (Poro['j'] == KG_50_Grids_hist["KG-50I_J"][i])]["Poro"])]
        
            
    for i in range(0,len(KG_56_Grids_hist.index)):
        perm_KG_56=np.r_[perm_KG_56,np.array(Real1.loc[(Real1['i'] == KG_56_Grids_hist["KG-56I_I"][i]) & (Real1['j'] == KG_56_Grids_hist["KG-56I_J"][i])][str(name)])]
        perm_poro_KG_56=np.r_[perm_poro_KG_56,np.array(Real1.loc[(Real1['i'] == KG_56_Grids_hist["KG-56I_I"][i]) & (Real1['j'] == KG_56_Grids_hist["KG-56I_J"][i])][str(name)])/np.array(Poro.loc[(Poro['i'] == KG_56_Grids_hist["KG-56I_I"][i]) & (Poro['j'] == KG_56_Grids_hist["KG-56I_J"][i])]["Poro"])]

            
    for i in range(0,len(KG_58_Grids_hist.index)):
        perm_KG_58=np.r_[perm_KG_58,np.array(Real1.loc[(Real1['i'] == KG_58_Grids_hist["KG-58I_I"][i]) & (Real1['j'] == KG_58_Grids_hist["KG-58I_J"][i])][str(name)])]
        perm_poro_KG_58=np.r_[perm_poro_KG_58,np.array(Real1.loc[(Real1['i'] == KG_58_Grids_hist["KG-58I_I"][i]) & (Real1['j'] == KG_58_Grids_hist["KG-58I_J"][i])][str(name)])/np.array(Poro.loc[(Poro['i'] == KG_58_Grids_hist["KG-58I_I"][i]) & (Poro['j'] == KG_58_Grids_hist["KG-58I_J"][i])]["Poro"])]

    for i in range(0,len(KG_50_02_Grids_hist.index)):
        perm_KG_50_02=np.r_[perm_KG_50_02,np.array(Real1.loc[(Real1['i'] == KG_50_02_Grids_hist["KG-50-02_I"][i]) & (Real1['j'] == KG_50_02_Grids_hist["KG-50-02_J"][i])][str(name)])]
        perm_poro_KG_50_02=np.r_[perm_poro_KG_50_02,np.array(Real1.loc[(Real1['i'] == KG_50_02_Grids_hist["KG-50-02_I"][i]) & (Real1['j'] == KG_50_02_Grids_hist["KG-50-02_J"][i])][str(name)])/np.array(Poro.loc[(Poro['i'] == KG_50_02_Grids_hist["KG-50-02_I"][i]) & (Poro['j'] == KG_50_02_Grids_hist["KG-50-02_J"][i])]["Poro"])]

    for i in range(0,len(KG_56_02_Grids_hist.index)):
        perm_KG_56_02=np.r_[perm_KG_56_02,np.array(Real1.loc[(Real1['i'] == KG_56_02_Grids_hist["KG-56-02_I"][i]) & (Real1['j'] == KG_56_02_Grids_hist["KG-56-02_J"][i])][str(name)])]
        perm_poro_KG_56_02=np.r_[perm_poro_KG_56_02,np.array(Real1.loc[(Real1['i'] == KG_56_02_Grids_hist["KG-56-02_I"][i]) & (Real1['j'] == KG_56_02_Grids_hist["KG-56-02_J"][i])][str(name)])/np.array(Poro.loc[(Poro['i'] == KG_56_02_Grids_hist["KG-56-02_I"][i]) & (Poro['j'] == KG_56_02_Grids_hist["KG-56-02_J"][i])]["Poro"])]

 
    KG_02_Ave_Perm=np.average(perm_KG_02)
    KG_04_Ave_Perm=np.average(perm_KG_04)
    KG_50_Ave_Perm=np.average(perm_KG_50)
    KG_56_Ave_Perm=np.average(perm_KG_56)
    KG_58_Ave_Perm=np.average(perm_KG_58)
    KG_50_02_Ave_Perm=np.average(perm_KG_50_02)
    KG_56_02_Ave_Perm=np.average(perm_KG_56_02)

    
    KG_02_Ave_Perm_Poro=np.average(perm_poro_KG_02)
    KG_04_Ave_Perm_Poro=np.average(perm_poro_KG_04)
    KG_50_Ave_Perm_Poro=np.average(perm_poro_KG_50)
    KG_56_Ave_Perm_Poro=np.average(perm_poro_KG_56)
    KG_58_Ave_Perm_Poro=np.average(perm_poro_KG_58)
    KG_50_02_Ave_Perm_Poro=np.average(perm_poro_KG_50_02)
    KG_56_02_Ave_Perm_Poro=np.average(perm_poro_KG_56_02)
    
    
    Ave_Perm_KG_02.append(KG_02_Ave_Perm)
    Ave_Perm_KG_04.append(KG_04_Ave_Perm)
    Ave_Perm_KG_50.append(KG_50_Ave_Perm)
    Ave_Perm_KG_56.append(KG_56_Ave_Perm)
    Ave_Perm_KG_58.append(KG_58_Ave_Perm)
    Ave_Perm_KG_50_02.append(KG_50_02_Ave_Perm)
    Ave_Perm_KG_56_02.append(KG_56_02_Ave_Perm)
    
    
    Ave_Perm_Phi_KG_02.append(KG_02_Ave_Perm_Poro)
    Ave_Perm_Phi_KG_04.append(KG_04_Ave_Perm_Poro)
    Ave_Perm_Phi_KG_50.append(KG_50_Ave_Perm_Poro)
    Ave_Perm_Phi_KG_56.append(KG_56_Ave_Perm_Poro)
    Ave_Perm_Phi_KG_58.append(KG_58_Ave_Perm_Poro)  
    Ave_Perm_Phi_KG_50_02.append(KG_50_02_Ave_Perm_Poro)
    Ave_Perm_Phi_KG_56_02.append(KG_56_02_Ave_Perm_Poro)

    
    KG_02_std_Perm=np.std(perm_KG_02)
    KG_04_std_Perm=np.std(perm_KG_04)
    KG_50_std_Perm=np.std(perm_KG_50)
    KG_56_std_Perm=np.std(perm_KG_56)
    KG_58_std_Perm=np.std(perm_KG_58)
    KG_50_02_std_Perm=np.std(perm_KG_50_02)
    KG_56_02_std_Perm=np.std(perm_KG_56_02)

    
    KG_02_std_Perm_Poro=np.std(perm_poro_KG_02)
    KG_04_std_Perm_Poro=np.std(perm_poro_KG_04)
    KG_50_std_Perm_Poro=np.std(perm_poro_KG_50)
    KG_56_std_Perm_Poro=np.std(perm_poro_KG_56)
    KG_58_std_Perm_Poro=np.std(perm_poro_KG_58)
    KG_50_02_std_Perm_Poro=np.std(perm_poro_KG_50_02)
    KG_56_02_std_Perm_Poro=np.std(perm_poro_KG_56_02)

    
    Ave_STD_KG_02.append(KG_02_std_Perm)
    Ave_STD_KG_04.append(KG_04_std_Perm)
    Ave_STD_KG_50.append(KG_50_std_Perm)
    Ave_STD_KG_56.append(KG_56_std_Perm)
    Ave_STD_KG_58.append(KG_58_std_Perm)
    Ave_STD_KG_50_02.append(KG_50_02_std_Perm)
    Ave_STD_KG_56_02.append(KG_56_02_std_Perm)
    
    
    Ave_STD_Phi_KG_02.append(KG_02_std_Perm_Poro)
    Ave_STD_Phi_KG_04.append(KG_04_std_Perm_Poro)
    Ave_STD_Phi_KG_50.append(KG_50_std_Perm_Poro)
    Ave_STD_Phi_KG_56.append(KG_56_std_Perm_Poro)
    Ave_STD_Phi_KG_58.append(KG_58_std_Perm_Poro)
    Ave_STD_Phi_KG_50_02.append(KG_50_02_std_Perm_Poro)
    Ave_STD_Phi_KG_56_02.append(KG_56_02_std_Perm_Poro)
    
    
np.savetxt("Ave_Perm_KG_02_I1.csv" , Ave_Perm_KG_02 , delimiter =",", fmt ='% s')
np.savetxt("Ave_Perm_KG_04_I1.csv" , Ave_Perm_KG_04 , delimiter =",", fmt ='% s')
np.savetxt("Ave_Perm_KG_50_I1.csv" , Ave_Perm_KG_50 , delimiter =",", fmt ='% s')
np.savetxt("Ave_Perm_KG_56_I1.csv" , Ave_Perm_KG_56 , delimiter =",", fmt ='% s')
np.savetxt("Ave_Perm_KG_58_I1.csv" , Ave_Perm_KG_58 , delimiter =",", fmt ='% s')
np.savetxt("Ave_Perm_KG_50_02_I1.csv" , Ave_Perm_KG_50_02 , delimiter =",", fmt ='% s')
np.savetxt("Ave_Perm_KG_56_02_I1.csv" , Ave_Perm_KG_56_02 , delimiter =",", fmt ='% s')


np.savetxt("Ave_Perm_Phi_KG_02_I1.csv" , Ave_Perm_Phi_KG_02 , delimiter =",", fmt ='% s')
np.savetxt("Ave_Perm_Phi_KG_04_I1.csv" , Ave_Perm_Phi_KG_04 , delimiter =",", fmt ='% s')
np.savetxt("Ave_Perm_Phi_KG_50_I1.csv" , Ave_Perm_Phi_KG_50 , delimiter =",", fmt ='% s')
np.savetxt("Ave_Perm_Phi_KG_56_I1.csv" , Ave_Perm_Phi_KG_56 , delimiter =",", fmt ='% s')
np.savetxt("Ave_Perm_Phi_KG_58_I1.csv" , Ave_Perm_Phi_KG_58 , delimiter =",", fmt ='% s')
np.savetxt("Ave_Perm_Phi_KG_50_02_I1.csv" , Ave_Perm_Phi_KG_50_02 , delimiter =",", fmt ='% s')
np.savetxt("Ave_Perm_Phi_KG_56_02_I1.csv" , Ave_Perm_Phi_KG_56_02 , delimiter =",", fmt ='% s')

    
np.savetxt("STD_Perm_KG_02_I1.csv" , Ave_STD_KG_02 , delimiter =",", fmt ='% s')
np.savetxt("STD_Perm_KG_04_I1.csv" , Ave_STD_KG_04 , delimiter =",", fmt ='% s')
np.savetxt("STD_Perm_KG_50_I1.csv" , Ave_STD_KG_50 , delimiter =",", fmt ='% s')  
np.savetxt("STD_Perm_KG_56_I1.csv" , Ave_STD_KG_56 , delimiter =",", fmt ='% s')
np.savetxt("STD_Perm_KG_58_I1.csv" , Ave_STD_KG_58 , delimiter =",", fmt ='% s')
np.savetxt("STD_Perm_KG_50_02_I1.csv" , Ave_STD_KG_50_02 , delimiter =",", fmt ='% s')  
np.savetxt("STD_Perm_KG_56_02_I1.csv" , Ave_STD_KG_56_02 , delimiter =",", fmt ='% s')


np.savetxt("STD_Perm_Phi_KG_02_I1.csv" , Ave_STD_Phi_KG_02 , delimiter =",", fmt ='% s')
np.savetxt("STD_Perm_Phi_KG_04_I1.csv" , Ave_STD_Phi_KG_04 , delimiter =",", fmt ='% s')
np.savetxt("STD_Perm_Phi_KG_50_I1.csv" , Ave_STD_Phi_KG_50 , delimiter =",", fmt ='% s')
np.savetxt("STD_Perm_Phi_KG_56_I1.csv" , Ave_STD_Phi_KG_56 , delimiter =",", fmt ='% s')
np.savetxt("STD_Perm_Phi_KG_58_I1.csv" , Ave_STD_Phi_KG_58 , delimiter =",", fmt ='% s')
np.savetxt("STD_Perm_Phi_KG_50_02_I1.csv" , Ave_STD_Phi_KG_50_02 , delimiter =",", fmt ='% s')
np.savetxt("STD_Perm_Phi_KG_56_02_I1.csv" , Ave_STD_Phi_KG_56_02 , delimiter =",", fmt ='% s')


ARMSE_KG_02=[]
ARMSE_KG_04=[]
ARMSE_KG_50=[]
ARMSE_KG_56=[]
ARMSE_KG_58=[]
ARMSE_KG_50_02=[]
ARMSE_KG_56_02=[]

RMSE_std_KG_02=[]
RMSE_std_KG_04=[]
RMSE_std_KG_50=[]
RMSE_std_KG_56=[]
RMSE_std_KG_58=[]
RMSE_std_KG_50_02=[]
RMSE_std_KG_56_02=[]

ARMSE_Phi_KG_02=[]
ARMSE_Phi_KG_04=[]
ARMSE_Phi_KG_50=[]
ARMSE_Phi_KG_56=[]
ARMSE_Phi_KG_58=[]
ARMSE_Phi_KG_50_02=[]
ARMSE_Phi_KG_56_02=[]


RMSE_std_Phi_KG_02=[]
RMSE_std_Phi_KG_04=[]
RMSE_std_Phi_KG_50=[]
RMSE_std_Phi_KG_56=[]
RMSE_std_Phi_KG_58=[]
RMSE_std_Phi_KG_50_02=[]
RMSE_std_Phi_KG_56_02=[]

Poro=pd.read_csv("sector_porosity.csv",header=0, delimiter=",", index_col=None)
Perm=pd.read_csv("sector_perm.csv",header=0, delimiter=",", index_col=None)




for name in range(1,401):
    print(name)
    
    Real1=Perm[["i","j","k",str(name)]]
    #######################################################################################################
    RMSE_KG_02=np.array([])
    RMSE_Phi_KG_02=np.array([])
    KG_02_Grids_hist_mat=np.array(KG_02_Grids_hist[["KG-02I_I","KG-02I_J"]])
    for items in KG_02_Grids_hist_mat:
        mse1=0
        mse2=0
        mse_phi1=0
        mse_phi2=0
        n=0
        a=np.array(Real1.loc[(Real1['i'] == items[0]) & (Real1['j'] == items[1])][str(name)])
        a_phi=np.array(Poro.loc[(Real1['i'] == items[0]) & (Poro['j'] == items[1])]["Poro"])
        if KG_02_Grids_hist_mat[(KG_02_Grids_hist_mat[:,0]==items[0]+1) & (KG_02_Grids_hist_mat[:,1]==items[1])].size!=0:
            b=np.array(Real1.loc[(Real1['i'] == items[0]+1) & (Real1['j'] == items[1])][str(name)])
            b_phi=np.array(Poro.loc[(Poro['i'] == items[0]+1) & (Poro['j'] == items[1])]["Poro"])
            mse1=np.average((a-b)**2)
            mse_phi1=np.average(((a/(a_phi+0.0000001))-(b/(b_phi+0.0000001)))**2)
            n+=1
        if KG_02_Grids_hist_mat[(KG_02_Grids_hist_mat[:,0]==items[0]) & (KG_02_Grids_hist_mat[:,1]==items[1]+1)].size!=0:
            c=np.array(Real1.loc[(Real1['i'] == items[0]) & (Real1['j'] == items[1]+1)][str(name)])
            c_phi=np.array(Poro.loc[(Real1['i'] == items[0]) & (Poro['j'] == items[1]+1)]["Poro"])
            mse2=np.average((a-c)**2)
            mse_phi2=np.average(((a/(a_phi+0.0000001))-(c/(c_phi+0.0000001)))**2)
            n+=1
        if n!=0:
            RMSE_KG_02=np.r_[RMSE_KG_02,np.sqrt((mse1+mse2)/n)]
            RMSE_Phi_KG_02=np.r_[RMSE_Phi_KG_02,np.sqrt((mse_phi1+mse_phi2)/n)]
        else:
            RMSE_KG_02=np.r_[RMSE_KG_02,[0]]
            RMSE_Phi_KG_02=np.r_[RMSE_Phi_KG_02,[0]]
    ARMSE1=np.average(RMSE_KG_02)
    ARMSE_Phi1=np.average(RMSE_Phi_KG_02)
    
    RMSE_KG_02_std=np.std(RMSE_KG_02)
    RMSE_KG_02_std_Phi=np.std(RMSE_Phi_KG_02)
    
    ARMSE_KG_02.append(ARMSE1)
    ARMSE_Phi_KG_02.append(ARMSE_Phi1)
    
    RMSE_std_KG_02.append(RMSE_KG_02_std)
    RMSE_std_Phi_KG_02.append(RMSE_KG_02_std_Phi)
    #########################################################################    
    RMSE_KG_04=np.array([])
    RMSE_Phi_KG_04=np.array([])
    KG_04_Grids_hist_mat=np.array(KG_04_Grids_hist[["KG-04I_I","KG-04I_J"]])
    for items in KG_04_Grids_hist_mat:
        mse1=0
        mse2=0
        mse_phi1=0
        mse_phi2=0
        n=0
        a=np.array(Real1.loc[(Real1['i'] == items[0]) & (Real1['j'] == items[1])][str(name)])
        a_phi=np.array(Poro.loc[(Real1['i'] == items[0]) & (Poro['j'] == items[1])]["Poro"])
        if KG_04_Grids_hist_mat[(KG_04_Grids_hist_mat[:,0]==items[0]+1) & (KG_04_Grids_hist_mat[:,1]==items[1])].size!=0:
            b=np.array(Real1.loc[(Real1['i'] == items[0]+1) & (Real1['j'] == items[1])][str(name)])
            b_phi=np.array(Poro.loc[(Poro['i'] == items[0]+1) & (Poro['j'] == items[1])]["Poro"])
            mse1=np.average((a-b)**2)
            mse_phi1=np.average(((a/(a_phi+0.0000001))-(b/(b_phi+0.0000001)))**2)
            n+=1
        if KG_04_Grids_hist_mat[(KG_04_Grids_hist_mat[:,0]==items[0]) & (KG_04_Grids_hist_mat[:,1]==items[1]+1)].size!=0:
            c=np.array(Real1.loc[(Real1['i'] == items[0]) & (Real1['j'] == items[1]+1)][str(name)])
            c_phi=np.array(Poro.loc[(Real1['i'] == items[0]) & (Poro['j'] == items[1]+1)]["Poro"])
            mse2=np.average((a-c)**2)
            mse_phi2=np.average(((a/(a_phi+0.0000001))-(c/(c_phi+0.0000001)))**2)
            n+=1
        if n!=0:
            RMSE_KG_04=np.r_[RMSE_KG_04,np.sqrt((mse1+mse2)/n)]
            RMSE_Phi_KG_04=np.r_[RMSE_Phi_KG_04,np.sqrt((mse_phi1+mse_phi2)/n)]
        else:
            RMSE_KG_04=np.r_[RMSE_KG_04,[0]]
            RMSE_Phi_KG_04=np.r_[RMSE_Phi_KG_04,[0]]
    ARMSE2=np.average(RMSE_KG_04)
    ARMSE_Phi2=np.average(RMSE_Phi_KG_04)
    
    RMSE_KG_04_std=np.std(RMSE_KG_04)
    RMSE_KG_04_std_Phi=np.std(RMSE_Phi_KG_04)
    
    ARMSE_KG_04.append(ARMSE2)
    ARMSE_Phi_KG_04.append(ARMSE_Phi2)
    
    RMSE_std_KG_04.append(RMSE_KG_04_std)
    RMSE_std_Phi_KG_04.append(RMSE_KG_04_std_Phi)
    ############################################################################## 
    RMSE_KG_50=np.array([])
    RMSE_Phi_KG_50=np.array([])
    KG_50_Grids_hist_mat=np.array(KG_50_Grids_hist[["KG-50I_I","KG-50I_J"]])
    for items in KG_50_Grids_hist_mat:
        mse1=0
        mse2=0
        mse_phi1=0
        mse_phi2=0
        n=0
        a=np.array(Real1.loc[(Real1['i'] == items[0]) & (Real1['j'] == items[1])][str(name)])
        a_phi=np.array(Poro.loc[(Real1['i'] == items[0]) & (Poro['j'] == items[1])]["Poro"])
        if KG_50_Grids_hist_mat[(KG_50_Grids_hist_mat[:,0]==items[0]+1) & (KG_50_Grids_hist_mat[:,1]==items[1])].size!=0:
            b=np.array(Real1.loc[(Real1['i'] == items[0]+1) & (Real1['j'] == items[1])][str(name)])
            b_phi=np.array(Poro.loc[(Poro['i'] == items[0]+1) & (Poro['j'] == items[1])]["Poro"])
            mse1=np.average((a-b)**2)
            mse_phi1=np.average(((a/(a_phi+0.0000001))-(b/(b_phi+0.0000001)))**2)
            n+=1
        if KG_50_Grids_hist_mat[(KG_50_Grids_hist_mat[:,0]==items[0]) & (KG_50_Grids_hist_mat[:,1]==items[1]+1)].size!=0:
            c=np.array(Real1.loc[(Real1['i'] == items[0]) & (Real1['j'] == items[1]+1)][str(name)])
            c_phi=np.array(Poro.loc[(Real1['i'] == items[0]) & (Poro['j'] == items[1]+1)]["Poro"])
            mse2=np.average((a-c)**2)
            mse_phi2=np.average(((a/(a_phi+0.0000001))-(c/(c_phi+0.0000001)))**2)
            n+=1
        if n!=0:
            RMSE_KG_50=np.r_[RMSE_KG_50,np.sqrt((mse1+mse2)/n)]
            RMSE_Phi_KG_50=np.r_[RMSE_Phi_KG_50,np.sqrt((mse_phi1+mse_phi2)/n)]
        else:
            RMSE_KG_50=np.r_[RMSE_KG_50,[0]]
            RMSE_Phi_KG_50=np.r_[RMSE_Phi_KG_50,[0]]
    ARMSE3=np.average(RMSE_KG_50)
    ARMSE_Phi3=np.average(RMSE_Phi_KG_50)
    
    RMSE_KG_50_std=np.std(RMSE_KG_50)
    RMSE_KG_50_std_Phi=np.std(RMSE_Phi_KG_50)
    
    ARMSE_KG_50.append(ARMSE3)
    ARMSE_Phi_KG_50.append(ARMSE3)
    
    RMSE_std_KG_50.append(RMSE_KG_50_std)
    RMSE_std_Phi_KG_50.append(RMSE_KG_50_std_Phi)    
    ############################################################################
    ############################################################################## 
    RMSE_KG_56=np.array([])
    RMSE_Phi_KG_56=np.array([])
    KG_56_Grids_hist_mat=np.array(KG_56_Grids_hist[["KG-56I_I","KG-56I_J"]])
    for items in KG_56_Grids_hist_mat:
        mse1=0
        mse2=0
        mse_phi1=0
        mse_phi2=0
        n=0
        a=np.array(Real1.loc[(Real1['i'] == items[0]) & (Real1['j'] == items[1])][str(name)])
        a_phi=np.array(Poro.loc[(Real1['i'] == items[0]) & (Poro['j'] == items[1])]["Poro"])
        if KG_56_Grids_hist_mat[(KG_56_Grids_hist_mat[:,0]==items[0]+1) & (KG_56_Grids_hist_mat[:,1]==items[1])].size!=0:
            b=np.array(Real1.loc[(Real1['i'] == items[0]+1) & (Real1['j'] == items[1])][str(name)])
            b_phi=np.array(Poro.loc[(Poro['i'] == items[0]+1) & (Poro['j'] == items[1])]["Poro"])
            mse1=np.average((a-b)**2)
            mse_phi1=np.average(((a/(a_phi+0.0000001))-(b/(b_phi+0.0000001)))**2)
            n+=1
        if KG_56_Grids_hist_mat[(KG_56_Grids_hist_mat[:,0]==items[0]) & (KG_56_Grids_hist_mat[:,1]==items[1]+1)].size!=0:
            c=np.array(Real1.loc[(Real1['i'] == items[0]) & (Real1['j'] == items[1]+1)][str(name)])
            c_phi=np.array(Poro.loc[(Real1['i'] == items[0]) & (Poro['j'] == items[1]+1)]["Poro"])
            mse2=np.average((a-c)**2)
            mse_phi2=np.average(((a/(a_phi+0.0000001))-(c/(c_phi+0.0000001)))**2)
            n+=1
        if n!=0:
            RMSE_KG_56=np.r_[RMSE_KG_56,np.sqrt((mse1+mse2)/n)]
            RMSE_Phi_KG_56=np.r_[RMSE_Phi_KG_56,np.sqrt((mse_phi1+mse_phi2)/n)]
        else:
            RMSE_KG_56=np.r_[RMSE_KG_56,[0]]
            RMSE_Phi_KG_56=np.r_[RMSE_Phi_KG_56,[0]]
    ARMSE4=np.average(RMSE_KG_56)
    ARMSE_Phi4=np.average(RMSE_Phi_KG_56)
    
    RMSE_KG_56_std=np.std(RMSE_KG_56)
    RMSE_KG_56_std_Phi=np.std(RMSE_Phi_KG_56)
    
    ARMSE_KG_56.append(ARMSE4)
    ARMSE_Phi_KG_56.append(ARMSE_Phi4)
    
    RMSE_std_KG_56.append(RMSE_KG_56_std)
    RMSE_std_Phi_KG_56.append(RMSE_KG_56_std_Phi)    
    ############################################################################
    ############################################################################## 
    RMSE_KG_58=np.array([])
    RMSE_Phi_KG_58=np.array([])
    KG_58_Grids_hist_mat=np.array(KG_58_Grids_hist[["KG-58I_I","KG-58I_J"]])
    for items in KG_58_Grids_hist_mat:
        mse1=0
        mse2=0
        mse_phi1=0
        mse_phi2=0
        n=0
        a=np.array(Real1.loc[(Real1['i'] == items[0]) & (Real1['j'] == items[1])][str(name)])
        a_phi=np.array(Poro.loc[(Real1['i'] == items[0]) & (Poro['j'] == items[1])]["Poro"])
        if KG_58_Grids_hist_mat[(KG_58_Grids_hist_mat[:,0]==items[0]+1) & (KG_58_Grids_hist_mat[:,1]==items[1])].size!=0:
            b=np.array(Real1.loc[(Real1['i'] == items[0]+1) & (Real1['j'] == items[1])][str(name)])
            b_phi=np.array(Poro.loc[(Poro['i'] == items[0]+1) & (Poro['j'] == items[1])]["Poro"])
            mse1=np.average((a-b)**2)
            mse_phi1=np.average(((a/(a_phi+0.0000001))-(b/(b_phi+0.0000001)))**2)
            n+=1
        if KG_58_Grids_hist_mat[(KG_58_Grids_hist_mat[:,0]==items[0]) & (KG_58_Grids_hist_mat[:,1]==items[1]+1)].size!=0:
            c=np.array(Real1.loc[(Real1['i'] == items[0]) & (Real1['j'] == items[1]+1)][str(name)])
            c_phi=np.array(Poro.loc[(Real1['i'] == items[0]) & (Poro['j'] == items[1]+1)]["Poro"])
            mse2=np.average((a-c)**2)
            mse_phi2=np.average(((a/(a_phi+0.0000001))-(c/(c_phi+0.0000001)))**2)
            n+=1
        if n!=0:
            RMSE_KG_58=np.r_[RMSE_KG_58,np.sqrt((mse1+mse2)/n)]
            RMSE_Phi_KG_58=np.r_[RMSE_Phi_KG_58,np.sqrt((mse_phi1+mse_phi2)/n)]
        else:
            RMSE_KG_58=np.r_[RMSE_KG_58,[0]]
            RMSE_Phi_KG_58=np.r_[RMSE_Phi_KG_58,[0]]
    ARMSE5=np.average(RMSE_KG_58)
    ARMSE_Phi5=np.average(RMSE_Phi_KG_58)
    
    RMSE_KG_58_std=np.std(RMSE_KG_58)
    RMSE_KG_58_std_Phi=np.std(RMSE_Phi_KG_58)
    
    ARMSE_KG_58.append(ARMSE5)
    ARMSE_Phi_KG_58.append(ARMSE5)
    
    RMSE_std_KG_58.append(RMSE_KG_58_std)
    RMSE_std_Phi_KG_58.append(RMSE_KG_58_std_Phi)    
    ############################################################################
    ############################################################################## 
    RMSE_KG_50_02=np.array([])
    RMSE_Phi_KG_50_02=np.array([])
    KG_50_02_Grids_hist_mat=np.array(KG_50_02_Grids_hist[["KG-50-02_I","KG-50-02_J"]])
    for items in KG_50_02_Grids_hist_mat:
        mse1=0
        mse2=0
        mse_phi1=0
        mse_phi2=0
        n=0
        a=np.array(Real1.loc[(Real1['i'] == items[0]) & (Real1['j'] == items[1])][str(name)])
        a_phi=np.array(Poro.loc[(Real1['i'] == items[0]) & (Poro['j'] == items[1])]["Poro"])
        if KG_50_02_Grids_hist_mat[(KG_50_02_Grids_hist_mat[:,0]==items[0]+1) & (KG_50_02_Grids_hist_mat[:,1]==items[1])].size!=0:
            b=np.array(Real1.loc[(Real1['i'] == items[0]+1) & (Real1['j'] == items[1])][str(name)])
            b_phi=np.array(Poro.loc[(Poro['i'] == items[0]+1) & (Poro['j'] == items[1])]["Poro"])
            mse1=np.average((a-b)**2)
            mse_phi1=np.average(((a/(a_phi+0.0000001))-(b/(b_phi+0.0000001)))**2)
            n+=1
        if KG_50_02_Grids_hist_mat[(KG_50_02_Grids_hist_mat[:,0]==items[0]) & (KG_50_02_Grids_hist_mat[:,1]==items[1]+1)].size!=0:
            c=np.array(Real1.loc[(Real1['i'] == items[0]) & (Real1['j'] == items[1]+1)][str(name)])
            c_phi=np.array(Poro.loc[(Real1['i'] == items[0]) & (Poro['j'] == items[1]+1)]["Poro"])
            mse2=np.average((a-c)**2)
            mse_phi2=np.average(((a/(a_phi+0.0000001))-(c/(c_phi+0.0000001)))**2)
            n+=1
        if n!=0:
            RMSE_KG_50_02=np.r_[RMSE_KG_50_02,np.sqrt((mse1+mse2)/n)]
            RMSE_Phi_KG_50_02=np.r_[RMSE_Phi_KG_50_02,np.sqrt((mse_phi1+mse_phi2)/n)]
        else:
            RMSE_KG_50_02=np.r_[RMSE_KG_50_02,[0]]
            RMSE_Phi_KG_50_02=np.r_[RMSE_Phi_KG_50_02,[0]]
    ARMSE6=np.average(RMSE_KG_50_02)
    ARMSE_Phi6=np.average(RMSE_Phi_KG_50_02)
    
    RMSE_KG_50_02_std=np.std(RMSE_KG_50_02)
    RMSE_KG_50_02_std_Phi=np.std(RMSE_Phi_KG_50_02)
    
    ARMSE_KG_50_02.append(ARMSE6)
    ARMSE_Phi_KG_50_02.append(ARMSE6)
    
    RMSE_std_KG_50_02.append(RMSE_KG_50_02_std)
    RMSE_std_Phi_KG_50_02.append(RMSE_KG_50_02_std_Phi)    
    ############################################################################
    ############################################################################## 
    RMSE_KG_56_02=np.array([])
    RMSE_Phi_KG_56_02=np.array([])
    KG_56_02_Grids_hist_mat=np.array(KG_56_02_Grids_hist[["KG-56-02_I","KG-56-02_J"]])
    for items in KG_56_02_Grids_hist_mat:
        mse1=0
        mse2=0
        mse_phi1=0
        mse_phi2=0
        n=0
        a=np.array(Real1.loc[(Real1['i'] == items[0]) & (Real1['j'] == items[1])][str(name)])
        a_phi=np.array(Poro.loc[(Real1['i'] == items[0]) & (Poro['j'] == items[1])]["Poro"])
        if KG_56_02_Grids_hist_mat[(KG_56_02_Grids_hist_mat[:,0]==items[0]+1) & (KG_56_02_Grids_hist_mat[:,1]==items[1])].size!=0:
            b=np.array(Real1.loc[(Real1['i'] == items[0]+1) & (Real1['j'] == items[1])][str(name)])
            b_phi=np.array(Poro.loc[(Poro['i'] == items[0]+1) & (Poro['j'] == items[1])]["Poro"])
            mse1=np.average((a-b)**2)
            mse_phi1=np.average(((a/(a_phi+0.0000001))-(b/(b_phi+0.0000001)))**2)
            n+=1
        if KG_56_02_Grids_hist_mat[(KG_56_02_Grids_hist_mat[:,0]==items[0]) & (KG_56_02_Grids_hist_mat[:,1]==items[1]+1)].size!=0:
            c=np.array(Real1.loc[(Real1['i'] == items[0]) & (Real1['j'] == items[1]+1)][str(name)])
            c_phi=np.array(Poro.loc[(Real1['i'] == items[0]) & (Poro['j'] == items[1]+1)]["Poro"])
            mse2=np.average((a-c)**2)
            mse_phi2=np.average(((a/(a_phi+0.0000001))-(c/(c_phi+0.0000001)))**2)
            n+=1
        if n!=0:
            RMSE_KG_56_02=np.r_[RMSE_KG_56_02,np.sqrt((mse1+mse2)/n)]
            RMSE_Phi_KG_56_02=np.r_[RMSE_Phi_KG_56_02,np.sqrt((mse_phi1+mse_phi2)/n)]
        else:
            RMSE_KG_56_02=np.r_[RMSE_KG_56_02,[0]]
            RMSE_Phi_KG_56_02=np.r_[RMSE_Phi_KG_56_02,[0]]
    ARMSE7=np.average(RMSE_KG_56_02)
    ARMSE_Phi7=np.average(RMSE_Phi_KG_56_02)
    
    RMSE_KG_56_02_std=np.std(RMSE_KG_56_02)
    RMSE_KG_56_02_std_Phi=np.std(RMSE_Phi_KG_56_02)
    
    ARMSE_KG_56_02.append(ARMSE7)
    ARMSE_Phi_KG_56_02.append(ARMSE_Phi7)
    
    RMSE_std_KG_56_02.append(RMSE_KG_56_02_std)
    RMSE_std_Phi_KG_56_02.append(RMSE_KG_56_02_std_Phi)    
    ############################################################################


np.savetxt("ARMSE_KG_02_I1.csv" , ARMSE_KG_02 , delimiter =",", fmt ='% s')
np.savetxt("ARMSE_KG_04_I1.csv" , ARMSE_KG_04 , delimiter =",", fmt ='% s')
np.savetxt("ARMSE_KG_50_I1.csv" , ARMSE_KG_50 , delimiter =",", fmt ='% s')
np.savetxt("ARMSE_KG_56_I1.csv" , ARMSE_KG_56 , delimiter =",", fmt ='% s')
np.savetxt("ARMSE_KG_58_I1.csv" , ARMSE_KG_58 , delimiter =",", fmt ='% s')
np.savetxt("ARMSE_KG_50_02_I1.csv" , ARMSE_KG_50_02 , delimiter =",", fmt ='% s')
np.savetxt("ARMSE_KG_56_02_I1.csv" , ARMSE_KG_56_02 , delimiter =",", fmt ='% s')


np.savetxt("RMSE_std_KG_02_I1.csv" , RMSE_std_KG_02 , delimiter =",", fmt ='% s')
np.savetxt("RMSE_std_KG_04_I1.csv" , RMSE_std_KG_04 , delimiter =",", fmt ='% s')
np.savetxt("RMSE_std_KG_50_I1.csv" , RMSE_std_KG_50 , delimiter =",", fmt ='% s')
np.savetxt("RMSE_std_KG_56_I1.csv" , RMSE_std_KG_56 , delimiter =",", fmt ='% s')
np.savetxt("RMSE_std_KG_58_I1.csv" , RMSE_std_KG_58 , delimiter =",", fmt ='% s')
np.savetxt("RMSE_std_KG_50_02_I1.csv" , RMSE_std_KG_50_02 , delimiter =",", fmt ='% s')
np.savetxt("RMSE_std_KG_56_02_I1.csv" , RMSE_std_KG_56_02 , delimiter =",", fmt ='% s')

np.savetxt("ARMSE_Phi_KG_02_I1.csv" , ARMSE_Phi_KG_02 , delimiter =",", fmt ='% s')
np.savetxt("ARMSE_Phi_KG_04_I1.csv" , ARMSE_Phi_KG_04 , delimiter =",", fmt ='% s')
np.savetxt("ARMSE_Phi_KG_50_I1.csv" , ARMSE_Phi_KG_50 , delimiter =",", fmt ='% s') 
np.savetxt("ARMSE_Phi_KG_56_I1.csv" , ARMSE_Phi_KG_56 , delimiter =",", fmt ='% s')
np.savetxt("ARMSE_Phi_KG_58_I1.csv" , ARMSE_Phi_KG_58 , delimiter =",", fmt ='% s')
np.savetxt("ARMSE_Phi_KG_50_02_I1.csv" , ARMSE_Phi_KG_50_02 , delimiter =",", fmt ='% s') 
np.savetxt("ARMSE_Phi_KG_56_02_I1.csv" , ARMSE_Phi_KG_56_02 , delimiter =",", fmt ='% s')

np.savetxt("RMSE_std_Phi_KG_02_I1.csv" , RMSE_std_Phi_KG_02 , delimiter =",", fmt ='% s')
np.savetxt("RMSE_std_Phi_KG_04_I1.csv" , RMSE_std_Phi_KG_04 , delimiter =",", fmt ='% s')
np.savetxt("RMSE_std_Phi_KG_50_I1.csv" , RMSE_std_Phi_KG_50 , delimiter =",", fmt ='% s')
np.savetxt("RMSE_std_Phi_KG_56_I1.csv" , RMSE_std_Phi_KG_56 , delimiter =",", fmt ='% s')
np.savetxt("RMSE_std_Phi_KG_58_I1.csv" , RMSE_std_Phi_KG_58 , delimiter =",", fmt ='% s')
np.savetxt("RMSE_std_Phi_KG_50_02_I1.csv" , RMSE_std_Phi_KG_50_02 , delimiter =",", fmt ='% s')
np.savetxt("RMSE_std_Phi_KG_56_02_I1.csv" , RMSE_std_Phi_KG_56_02 , delimiter =",", fmt ='% s')
