==> 10-Extracting_the_Field_Behaviors_from_RMS_files.ipynb <==
import pandas as pd
import numpy as np
Real_ML=["1","32","37","44","97","126","199","221","337","343","347","363","377","384","388"]
Days=list(pd.read_csv("Days", header=None)[0])


## Extracting the Wells Behavior
indx=[]
indx_BP9=[]
indx_GPR=[]
FOPT_KG=[]
FGPT_KG=[]
FPR_KG=[]

m=0
for item in Real_ML:
    name=[str(item),".RSM"]
    names="".join(name)
    a=pd.read_table(names, header=None)

    w=(a[a[4]=='WOPR        '].index.tolist())
    index1=[]
    for i in Days:
        index1.append(a[a[1]==str(i)].index.tolist())
    for j in range(0,len(Days)):
        indx.append(index1[j][16])
        indx_BP9.append(index1[j][18])
        indx_GPR.append(index1[j][17])   
    for k in indx[len(Days)*m:len(Days)*m+len(Days)]:
        FOPT_KG.append(float(a[4][k]))

    for q in indx_BP9[len(Days)*m:len(Days)*m+len(Days)]:
        FPR_KG.append(float(a[6][q]))

    for s in indx_GPR[len(Days)*m:len(Days)*m+len(Days)]:
        FGPT_KG.append(float(a[5][s]))

    m+=1
#WOPR1=np.reshape(WOPR1,(item+1,len(b)))
#WOPR2=np.reshape(WOPR2,(item+1,len(b)))
#WOPR3=np.reshape(WOPR3,(item+1,len(b)))
#WGPR1=np.reshape(WGPR1,(item+1,len(b)))
#WGPR2=np.reshape(WGPR2,(item+1,len(b)))
#WGPR3=np.reshape(WGPR3,(item+1,len(b)))
#WBP9_1=np.reshape(WOPR1,(item+1,len(b)))
#WBP9_2=np.reshape(WOPR2,(item+1,len(b)))
#WBP9_3=np.reshape(WOPR3,(item+1,len(b)))

FOPT_KG=np.transpose(np.reshape(FOPT_KG,(15,len(Days))))

FGPT_KG=np.transpose(np.reshape(FGPT_KG,(15,len(Days))))

FPR_KG=np.transpose(np.reshape(FPR_KG,(15,len(Days))))

np.savetxt("FOPT_KG.csv" , FOPT_KG , delimiter =", ", fmt ='% s') 

np.savetxt("FGPT_KG.csv" , FGPT_KG , delimiter =", ", fmt ='% s') 

np.savetxt("FPR_KG.csv" , FPR_KG , delimiter =", ", fmt ='% s')
